+++ tests/testthat/test_MsqlBackend.R
@@ -110,3 +110,108 @@ test_that("spectraNames,spectraNames<-,MsqlBackend", {
 
     expect_error(spectraNames(mm8_be) <- 1:20, "does not support")
 })
+
+test_that("filterMsLevel,MsqlBackend works", {
+    res <- filterMsLevel(mm8_be)
+    expect_equal(res, mm8_be)
+
+    res <- filterMsLevel(mm8_be, msLevel = 1:2)
+    expect_equal(res, mm8_be)
+
+    res <- filterMsLevel(mm8_be, msLevel = 3)
+    expect_true(length(res) == 0)
+
+    tmp <- mm8_be
+    tmp$msLevel <- rep(1:2, 99)
+    res <- filterMsLevel(tmp, msLevel = 1L)
+    expect_true(length(res) == (length(tmp) / 2))
+})
+
+test_that("filterRt,MsqlBackend works", {
+    res <- filterRt(mm8_be)
+    expect_equal(res, mm8_be)
+
+    res <- filterRt(mm8_be, rt = c(1000, 2000))
+    expect_true(length(res) == 0)
+
+    res <- filterRt(mm8_be, rt = c(10, 20))
+    expect_true(all(res$rtime > 10 & res$rtime < 20))
+
+    res <- filterRt(mm8_be, rt = c(10, 20), msLevel. = 2)
+    expect_equal(res, mm8_be)
+
+    tmp <- mm8_be
+    tmp$msLevel <- sample(1:3, length(tmp), replace = TRUE)
+    res <- filterRt(tmp, rt = c(10, 20), msLevel. = 3)
+    res_3 <- filterMsLevel(res, 3)
+    expect_true(all(rtime(res_3) >= 10 & rtime(res_3) <= 20))
+    expect_equal(filterMsLevel(res, c(1, 2)), filterMsLevel(tmp, c(1, 2)))
+
+    ## TMT
+    res <- filterRt(tmt_be, rt = c(200, 210), msLevel. = 2)
+    res_2 <- filterMsLevel(res, 2)
+    expect_true(all(rtime(res_2) >= 200 & rtime(res_2) <= 210))
+    expect_false(all(rtime(res) >= 200 & rtime(res) <= 210))
+    expect_equal(filterMsLevel(res, 1), filterMsLevel(tmt_be, 1))
+
+    res2 <- filterRt(res, rt = c(205, 210), msLevel. = 1)
+    res2_1 <- filterMsLevel(res2, 1)
+    expect_true(all(rtime(res2_1) >= 205 & rtime(res2_1) <= 210))
+    expect_equal(filterMsLevel(res2, 2), filterMsLevel(res, 2))
+})
+
+test_that("filterDataOrigin works", {
+    res <- filterDataOrigin(mm_be, normalizePath(mm8_file))
+    expect_true(all(res$dataOrigin == normalizePath(mm8_file)))
+
+    res <- filterDataOrigin(mm_be, normalizePath(mm14_file))
+    expect_true(all(res$dataOrigin == normalizePath(mm14_file)))
+
+    res <- filterDataOrigin(mm_be, normalizePath(c(mm14_file, mm8_file)))
+    expect_equal(unique(dataOrigin(res)), normalizePath(c(mm14_file, mm8_file)))
+})
+
+test_that("filterPrecursorMzRange works", {
+    res <- filterPrecursorMzRange(tmt_be, c(660, 670))
+    tmp <- tmt_be
+    tmp$precursorMz <- precursorMz(tmt_be)
+    res_2 <- filterPrecursorMzRange(tmp, c(660, 670))
+    expect_equal(peaksData(res), peaksData(res_2))
+    expect_true(all(precursorMz(res) >= 660 & precursorMz(res) <= 670))
+
+    expect_equal(filterPrecursorMzRange(tmt_be), tmt_be)
+})
+
+test_that("filterPrecursorMzValues works", {
+    tmt_be2 <- tmt_be
+    tmt_be2$precursorMz <- tmt_be$precursorMz
+    pmz <- c(620.1, 404.25, 417.7, 506.6)
+
+    res <- filterPrecursorMzValues(tmt_be, pmz, tolerance = 0.1)
+    res_2 <- filterPrecursorMzValues(tmt_be2, pmz, tolerance = 0.1)
+    expect_equal(length(res), length(res_2))
+    expect_equal(precursorMz(res), precursorMz(res_2))
+
+    res_3 <- filterPrecursorMzValues(tmt_mzr, pmz, tolerance = 0.1)
+    expect_equal(length(res), length(res_3))
+    expect_equal(precursorMz(res), precursorMz(res_3))
+
+    res <- filterPrecursorMzValues(tmt_be, sort(pmz), tolerance = 0.1)
+    res_2 <- filterPrecursorMzValues(tmt_be2, sort(pmz), tolerance = 0.1)
+    res_3 <- filterPrecursorMzValues(tmt_mzr, sort(pmz), tolerance = 0.1)
+    expect_equal(precursorMz(res), precursorMz(res_2))
+    expect_equal(precursorMz(res), precursorMz(res_3))
+
+    pmz <- c(456.3, 503.7815)
+    res <- filterPrecursorMzValues(tmt_be, pmz)
+    res_2 <- filterPrecursorMzValues(tmt_be2, pmz)
+    res_3 <- filterPrecursorMzValues(tmt_mzr, pmz)
+    expect_equal(precursorMz(res), precursorMz(res_2))
+    expect_equal(precursorMz(res), precursorMz(res_3))
+
+    res <- filterPrecursorMzValues(tmt_be, pmz[c(2, 1)])
+    res_2 <- filterPrecursorMzValues(tmt_be2, pmz[c(2, 1)])
+    res_3 <- filterPrecursorMzValues(tmt_mzr, pmz[c(2, 1)])
+    expect_equal(precursorMz(res), precursorMz(res_2))
+    expect_equal(precursorMz(res), precursorMz(res_3))
+})